requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
bokeh==2.0.2
bson==0.5.8
cycler==0.10.0
dataclasses==0.6
Django==2.2.12
django-bootstrap4==1.1.1
django-chartit==0.2.9
djongo==1.3.2
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.4
packaging==20.3
pandas==1.0.3
Pillow==7.1.2
plotly==4.7.0
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
retrying==1.3.3
six==1.14.0
soupsieve==2.0
sqlparse==0.2.4
tornado==6.0.4
typing-extensions==3.7.4.2
wordcloud==1.7.0
